% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk_90_10.R
\name{xwalk_90_10}
\alias{xwalk_90_10}
\title{Crosswalk tracts from 1990 to 2010 boundaries}
\usage{
xwalk_90_10(census_data, tractID)
}
\arguments{
\item{census_data}{An R object containing ACS/Census data w/ 1990 boundaries}

\item{tractID}{variable name that contains the tract ID}
}
\value{
Crosswalked ACS/Census data (2010 boundaries)
}
\description{
This function takes ACS/Census data with 1990 tract boundaries, and loads in
weights from the LTBD
(https://s4.ad.brown.edu/projects/diversity/Researcher/LTBDDload/DataList.aspx).
It assumes that all numeric values in the specified ACS/Census data are to be
crosswalked. If this is not the case, those variables should be filtered
prior to using the function.
It outputs the original data crosswalked to 2010 boundaries.
}
