% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk_00_10.R
\name{xwalk_00_10}
\alias{xwalk_00_10}
\title{Crosswalk tracts from 2000 to 2010 boundaries}
\usage{
xwalk_00_10(census_data, tractID, vars = names(census_data))
}
\arguments{
\item{census_data}{An R table containing ACS/Census data w/ 2000 boundaries}

\item{tractID}{Variable name that contains the tract ID}

\item{vars}{Census/ACS variables to crosswalk. Defaults to all numeric
variables, but can be adjusted. Custom variables must be a single string
or a list of strings.}
}
\value{
Crosswalked selected ACS/Census data (2010 boundaries, defaults to
all numeric data)
}
\description{
This function takes ACS/Census data with 2000 tract boundaries, and loads in
weights from the LTBD
(https://s4.ad.brown.edu/projects/diversity/Researcher/LTBDDload/DataList.aspx).
It assumes that you want all variables to be crosswalked, and includes a filter
to only choose numeric type variables. If you want to only crosswalk specific
variables, specify this in the vars argument (must be numeric).
It outputs the original data crosswalked to 2010 boundaries.
}
